LC-smallest-integer-divisible-by-k [Runtime: N/A; Memory: N/A]

+++ math/smallest-integer-divisible-by-k.cpp
@@ -4,7 +4,7 @@ public:
         if(k==1)return 1;
 
         if(k%2==0 || k%5==0)return -1;
-    int rem=0;
+    long long rem=0;
         for(int i=1;i<=k; i++){
             rem=((rem*10)+1);
             if(rem%k==0)return i;
